[refactor] password modal
